Move thread_pool and brdf to raylib
